refactor: Extract modifier functions to new file
Extract modifier-retrieving functions into `kolmafia-modifier.d.ts`.
This should make it easier to manage documentations of closely-related
functions (that share much of their internal behavior together).

Since this new classification is mostly arbitrary, other functions that
interact with modifiers (e.g. `modifierEval()`) have not been moved yet.
