+++ src/index.d.ts
@@ -1,2 +1,3 @@
 import './global';
 export * from './kolmafia';
+export * from './kolmafia-modifier';
